Дописаны ограничения

+++ src/problems/travelling_salesman/types.rs
@@ -1,4 +1,4 @@
 pub type City = usize;
 pub type Matrix = Vec<Vec<f64>>;
-pub type TimeMatrix = Vec<Vec<u32>>;
+pub type TimeMatrix = Vec<Vec<usize>>;
 pub type RuleStr = String;
